+++ src/app/[locale]/globals.scss
@@ -8,7 +8,6 @@
 @use '@/common/styles/buttons/closebutton.scss';
 @use '@/common/styles/inputs/searchinput.scss';
 @use '@/common/styles/inputs/textinput.scss';
-@use '@/common/styles/inputs/searchinput.scss';
 
 .container {
   background: #f6f6f6;
